+++ spec/spec_helper.rb
@@ -2,13 +2,29 @@
 require 'webmock'
 require 'webmock/rspec'
 require 'rspec/wait'
-require 'pry'
 require 'rack'
 require 'rack/test'
 require 'rake'
 
 require 'airbrake'
 require 'airbrake/rake/tasks'
+
+Airbrake.configure do |c|
+  c.project_id = 113743
+  c.project_key = 'fd04e13d806a90f96614ad8e529b2822'
+  c.logger = Logger.new('/dev/null')
+  c.app_version = '1.2.3'
+  c.workers = 5
+end
+
+RSpec.configure do |c|
+  c.order = 'random'
+  c.color = true
+  c.disable_monkey_patching!
+  c.wait_timeout = 3
+
+  c.include Rack::Test::Methods
+end
 
 # Load integration tests only when they're run through appraisals.
 if ENV['APPRAISAL_INITIALIZED']
@@ -73,23 +89,6 @@ if ENV['APPRAISAL_INITIALIZED']
   rescue LoadError
     puts '** Skipped Rack specs'
   end
-end
-
-RSpec.configure do |c|
-  c.order = 'random'
-  c.color = true
-  c.disable_monkey_patching!
-  c.wait_timeout = 3
-
-  c.include Rack::Test::Methods
-end
-
-Airbrake.configure do |c|
-  c.project_id = 113743
-  c.project_key = 'fd04e13d806a90f96614ad8e529b2822'
-  c.logger = Logger.new('/dev/null')
-  c.app_version = '1.2.3'
-  c.workers = 5
 end
 
 # Make sure tests that use async requests fail.